# api/openapi.yaml
openapi: 3.1.0
...
info:
  title: CameraAI Orchestrator API
  version: 0.1.0
  description: |
    API cua he thong camera giam sat hanh vi ung dung AI va AWS.

    ## Quy uoc chung
    - Tat ca endpoint (tru `/health`, `/auth/login` va `/auth/refresh`) yeu cau header
      `Authorization: Bearer <access_token>`.
    - Thoi gian luon la ISO-8601 co mui gio (`2026-09-15T08:30:00+07:00`).
    - Danh sach luon tra ve dang `{ data: [...], meta: { page, pageSize, total } }`.
    - Loi luon tra ve dang `{ error: { code, message, details?, traceId } }`.

    ## Trang thai su kien
    `DETECTED -> LOGGED_ONLY | NOTIFIED -> ESCALATED | RESOLVED -> CLOSED`
    (chi tiet: docs/architecture/C4_ARCHITECTURE.md, muc "State machine").
  contact:
    name: Nhom CameraAI
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:3001/api/v1
    description: Dev local (docker compose)
  - url: https://{host}/api/v1
    description: AWS (Sprint 4)
    variables:
      host:
        default: api.camerai.example.com
...
security:
  - bearerAuth: []
paths:
...
  /cameras/{cameraId}/zones:
    parameters:
      - $ref: "#/components/parameters/CameraId"
    get:
      tags: [zones]
      summary: Danh sach vung cua mot camera
      operationId: listZones
      responses:
        "200":
          description: Danh sach vung
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Zone" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [zones]
      summary: Tao vung giam sat moi
      description: |
        Toa do da giac phai CHUAN HOA ve khoang 0..1 de doc lap do phan giai.
        Orchestrator tu sinh lai cau hinh Frigate sau khi luu. Response thanh cong
        chi xac nhan zone da duoc luu vao PostgreSQL; trang thai `frigateSync` cua
        camera cho biet cau hinh da co hieu luc tren Frigate hay chua.
      operationId: createZone
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateZoneRequest" }
      responses:
        "201":
          description: Da tao
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Zone" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /zones/{zoneId}:
    parameters:
      - $ref: "#/components/parameters/ZoneId"
    patch:
      tags: [zones]
      summary: Cap nhat vung
      operationId: updateZone
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateZoneRequest" }
      responses:
        "200":
          description: Da cap nhat
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Zone" }
        "404": { $ref: "#/components/responses/NotFound" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      tags: [zones]
      summary: Xoa vung
      operationId: deleteZone
      responses:
        "204": { description: Da xoa }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ===================================================================
  # KNOWN FACES — US-09
  # ===================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Sprint 1-3: JWT do orchestrator tu cap.
        Sprint 4 (US-25): token do Cognito cap, xac thuc bang JWKS cua User Pool.
    internalToken:
      type: apiKey
      in: header
      name: X-Internal-Token
      description: Secret dung chung giua cac service noi bo, khong lo ra ngoai.
  parameters:
...
    CameraId:
      name: cameraId
      in: path
      required: true
      schema: { type: string, format: uuid }
    ZoneId:
      name: zoneId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Du lieu gui len khong hop le
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: Chua xac thuc hoac token het han (FR-AUT-06)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: Khong du quyen cho vai tro hien tai (FR-AUT-05)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Khong tim thay tai nguyen
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Conflict:
      description: Xung dot voi du lieu da ton tai
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
...
    ZoneType:
      type: string
      enum: [RESTRICTED, REST_AREA, NORMAL]
      description: REST_AREA (giuong/sofa) khong sinh canh bao te nga — FR-DET-M2-05.
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, traceId]
          properties:
            code:
              $ref: "#/components/schemas/CameraErrorCode"
            message:
              type: string
              description: Thong diep tieng Viet hien thi duoc cho nguoi dung.
            details:
              type: object
              additionalProperties: true
            traceId:
              type: string
              description: FR-LOG-02 — dung de tra log xuyen service.
    CameraErrorCode:
      type: string
      enum:
        - ACCOUNT_LOCKED
        - CAMERA_NOT_FOUND
        - CAMERA_SOURCE_NOT_FOUND
        - SOURCE_NOT_CONFIGURED
        - SOURCE_UNAVAILABLE
        - WEBCAM_PERMISSION_DENIED
        - WEBCAM_PUBLISH_FAILED
        - VIDEO_INVALID_FORMAT
        - VIDEO_TOO_LARGE
        - VIDEO_NOT_FOUND
        - FFMPEG_START_FAILED
        - FRIGATE_SYNC_FAILED
        - FRIGATE_UNAVAILABLE
        - MEDIAMTX_UNAVAILABLE
        - CAMERA_ALREADY_TRANSITIONING
        - CONFIG_VERSION_CONFLICT
        - FORBIDDEN
        - INVALID_CREDENTIALS
        - INVALID_REFRESH_TOKEN
        - INVALID_SLUG
        - INVALID_TOKEN
        - EVENT_NOT_FOUND
        - UNAUTHORIZED
      description: Stable errors returned by camera source, runtime and Frigate operations.
...
    Polygon:
      type: array
      minItems: 3
      description: Da giac voi toa do CHUAN HOA 0..1 theo [x, y].
      items:
        type: array
        minItems: 2
        maxItems: 2
        items: { type: number, minimum: 0, maximum: 1 }
      examples:
        - [[0.4, 0.3], [0.95, 0.3], [0.95, 0.9], [0.4, 0.9]]
    Zone:
      type: object
      required:
        [
          id,
          cameraId,
          name,
          slug,
          zoneType,
          polygon,
          minDwellSeconds,
          activeFrom,
          activeTo,
          isEnabled,
        ]
      properties:
        id: { type: string, format: uuid }
        cameraId: { type: string, format: uuid }
        name: { type: string, minLength: 1, maxLength: 60 }
        slug: { type: string }
        zoneType: { $ref: "#/components/schemas/ZoneType" }
        polygon: { $ref: "#/components/schemas/Polygon" }
        minDwellSeconds:
          type: integer
          minimum: 0
          maximum: 300
          description: FR-DET-M4-02 — di luot qua duoi nguong nay khong sinh canh bao.
        activeFrom:
          type: ["string", "null"]
          pattern: "^(?:[01][0-9]|2[0-3]):[0-5][0-9]$"
          examples: ["06:00"]
        activeTo:
          type: ["string", "null"]
          pattern: "^(?:[01][0-9]|2[0-3]):[0-5][0-9]$"
          examples: ["22:00"]
        isEnabled: { type: boolean }
    CreateZoneRequest:
      type: object
      required: [name, slug, zoneType, polygon]
      properties:
        name: { type: string, minLength: 1, maxLength: 60 }
        slug: { type: string, pattern: "^[a-z][a-z0-9_]{2,63}$" }
        zoneType: { $ref: "#/components/schemas/ZoneType" }
        polygon: { $ref: "#/components/schemas/Polygon" }
        minDwellSeconds: { type: integer, minimum: 0, maximum: 300, default: 2 }
        activeFrom:
          type: ["string", "null"]
          pattern: "^(?:[01][0-9]|2[0-3]):[0-5][0-9]$"
          description: Hai dau cung null la ca ngay; khoang [activeFrom, activeTo) co the qua nua dem.
        activeTo:
          type: ["string", "null"]
          pattern: "^(?:[01][0-9]|2[0-3]):[0-5][0-9]$"
          description: Khong duoc bang activeFrom.
    UpdateZoneRequest:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 60 }
        zoneType: { $ref: "#/components/schemas/ZoneType" }
        polygon: { $ref: "#/components/schemas/Polygon" }
        minDwellSeconds: { type: integer, minimum: 0, maximum: 300 }
        activeFrom:
          type: ["string", "null"]
          pattern: "^(?:[01][0-9]|2[0-3]):[0-5][0-9]$"
        activeTo:
          type: ["string", "null"]
          pattern: "^(?:[01][0-9]|2[0-3]):[0-5][0-9]$"
        isEnabled: { type: boolean }

    # ---------------- Known faces ----------------
...